move common/ package dependency from Utility to ELA
