+++ pact/protocol.go
@@ -1,0 +1,5 @@
+package pact
+
+const (
+	PrefixRegisterId = 0x67
+)
