+++ params/genesis.go
@@ -6,8 +6,9 @@ import (
 	"github.com/elastos/Elastos.ELA.SideChain/auxpow"
 	"github.com/elastos/Elastos.ELA.SideChain/types"
 
-	"github.com/elastos/Elastos.ELA.Utility/common"
-	"github.com/elastos/Elastos.ELA/core"
+	"github.com/elastos/Elastos.ELA/common"
+	ela "github.com/elastos/Elastos.ELA/core/types"
+	"github.com/elastos/Elastos.ELA/core/types/payload"
 )
 
 var (
@@ -38,16 +39,16 @@ var (
 		Version:    types.BlockVersion,
 		Previous:   common.EmptyHash,
 		MerkleRoot: ElaAssetId,
-		Timestamp:  uint32(time.Unix(time.Date(2018, time.June, 30,
+		Timestamp: uint32(time.Unix(time.Date(2018, time.June, 30,
 			12, 0, 0, 0, time.UTC).Unix(), 0).Unix()),
-		Bits:       0x1d03ffff,
-		Nonce:      types.GenesisNonce,
-		Height:     uint32(0),
+		Bits:   0x1d03ffff,
+		Nonce:  types.GenesisNonce,
+		Height: uint32(0),
 		SideAuxPow: auxpow.SideAuxPow{
-			SideAuxBlockTx: core.Transaction{
-				TxType:         core.SideChainPow,
-				PayloadVersion: core.SideChainPowPayloadVersion,
-				Payload:        new(core.PayloadSideChainPow),
+			SideAuxBlockTx: ela.Transaction{
+				TxType:         ela.SideChainPow,
+				PayloadVersion: payload.SideChainPowPayloadVersion,
+				Payload:        new(payload.PayloadSideChainPow),
 			},
 		},
 	}
